karibdocs/scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
langdetect>=1.0.9
pdfplumber>=0.11.0
pg8000==1.31.2